+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/TestHelpers.java
@@ -20,6 +20,7 @@
 
 package com.apple.foundationdb.record;
 
+import com.apple.foundationdb.record.provider.common.StoreTimer;
 import com.apple.foundationdb.record.provider.foundationdb.FDBRecordContext;
 import com.apple.foundationdb.record.provider.foundationdb.FDBStoreTimer;
 import com.apple.foundationdb.util.LoggableException;
@@ -68,27 +69,28 @@ public class TestHelpers {
         };
     }
 
-    public static void assertLogs(Class<?> loggingClass, Pattern pattern, Callable<?> callable) {
-        assertLogs(loggingClass.getName(), pattern, callable);
+    public static List<String> assertLogs(Class<?> loggingClass, Pattern pattern, Callable<?> callable) {
+        return assertLogs(loggingClass.getName(), pattern, callable);
     }
 
-    public static void assertLogs(String loggerName, Pattern pattern, Callable<?> callable) {
+    public static List<String> assertLogs(String loggerName, Pattern pattern, Callable<?> callable) {
         MatchingAppender appender = new MatchingAppender(UUID.randomUUID().toString(), pattern);
-        assertLogs(loggerName, appender, callable);
+        return assertLogs(loggerName, appender, callable);
     }
 
-    public static void assertLogs(Class<?> loggingClass, String messagePrefix, Callable<?> callable) {
-        assertLogs(loggingClass.getName(), messagePrefix, callable);
+    public static List<String> assertLogs(Class<?> loggingClass, String messagePrefix, Callable<?> callable) {
+        return assertLogs(loggingClass.getName(), messagePrefix, callable);
     }
 
-    public static void assertLogs(String loggerName, String messagePrefix, Callable<?> callable) {
+    public static List<String> assertLogs(String loggerName, String messagePrefix, Callable<?> callable) {
         MatchingAppender appender = new MatchingAppender(UUID.randomUUID().toString(), messagePrefix);
-        assertLogs(loggerName, appender, callable);
+        return assertLogs(loggerName, appender, callable);
     }
 
-    private static void assertLogs(String loggerName, MatchingAppender appender, Callable<?> callable) {
+    private static List<String> assertLogs(String loggerName, MatchingAppender appender, Callable<?> callable) {
         callAndMonitorLogging(loggerName, appender, callable);
         assertTrue(appender.matched(), () -> "No messages were logged matching [" + appender + "]");
+        return appender.getMatchedEvents();
     }
 
     public static void assertDidNotLog(Class<?> loggingClass, Pattern pattern, Callable<?> callable) {
@@ -112,6 +114,42 @@ public class TestHelpers {
     private static void assertDidNotLog(String loggerName, MatchingAppender appender, Callable<?> callable) {
         callAndMonitorLogging(loggerName, appender, callable);
         assertFalse(appender.matched(), () -> "Test should not have produced log message matching [" + appender + "]");
+    }
+
+    /**
+     * Returns a pattern that matches the count of an event in a log message. The pattern contains
+     * one group, which matches the count of the event in a log message.
+     *
+     * @param event store timer event to search for
+     * @return a {@link Pattern} that can be used to find event counts
+     */
+    public static Pattern eventCountPattern(StoreTimer.Event event) {
+        return Pattern.compile(".*" + event.logKeyWithSuffix("_count") + "=\"(\\d+)\".*");
+    }
+
+    /**
+     * Extract the count of an event from a log event. The supplied {@link Pattern} should come from
+     * {@link #eventCountPattern(StoreTimer.Event)} or should be equivalent.
+     *
+     * @param pattern pattern to use to extract a log event count
+     * @param logEvent a log formatted message string
+     * @return the count parsed from the log message
+     */
+    public static int extractCount(Pattern pattern, String logEvent) {
+        java.util.regex.Matcher matcher = pattern.matcher(logEvent);
+        assertTrue(matcher.matches(), () -> String.format("expected \"%s\" to match pattern \"%s\"", logEvent, pattern));
+        return Integer.parseInt(matcher.group(1));
+    }
+
+    /**
+     * Assert that a string does not match a pattern.
+     *
+     * @param pattern pattern to search for
+     * @param s string to search through
+     */
+    public static void assertDoesNotMatch(Pattern pattern, String s) {
+        java.util.regex.Matcher matcher = pattern.matcher(s);
+        assertFalse(matcher.matches(), () -> String.format("did not expect \"%s\" to have pattern \"%s\"", s, pattern));
     }
 
     private static void callAndMonitorLogging(String loggerName, MatchingAppender appender, Callable<?> callable) {
@@ -309,7 +347,7 @@ public class TestHelpers {
         @Nullable
         private final String messagePrefix;
 
-        private List<LogEvent> matchedEvents = new ArrayList<>();
+        private final List<String> matchedEvents = new ArrayList<>();
 
         protected MatchingAppender(@Nonnull String name, @Nonnull Pattern pattern) {
             super(name, null, null, true, null);
@@ -328,15 +366,15 @@ public class TestHelpers {
         }
 
         @Nonnull
-        public List<LogEvent> getMatchedEvents() {
+        public List<String> getMatchedEvents() {
             return matchedEvents;
         }
 
         @Override
-        public void append(@Nonnull LogEvent event) {
+        public synchronized void append(@Nonnull LogEvent event) {
             if ((pattern != null && pattern.matcher(event.getMessage().getFormattedMessage()).matches())
                     || (messagePrefix != null && event.getMessage().getFormattedMessage().startsWith(messagePrefix)))  {
-                matchedEvents.add(event);
+                matchedEvents.add(event.getMessage().getFormattedMessage());
             }
         }